file structure refactoring and vertical ipad responsiveness added
